committed fv_mp

+++ routes/web.php
@@ -11,107 +11,14 @@
 |
 */
 
-// Route::get('/', function () {
-//     return view('welcome');
-// });
-
-// Route::get('/index', function () {
-//     return view('public_website.index');
-// });
-
-
-// Route::get('/courses', function () {
-//     return view('public_website.courses_grid');
-// });
-
-
-Route::get('/singlecourse', function () {
-    return view('public_website.single_course');
-});
-
-
-// Route::get('/cart', function () {
-//     return view('public_website.cart');
-// });
-
-
-
-// Route::get('/checkout', function () {
-//     return view('public_website.checkout');
-// });
-
-Route::get('/dashboard', function () {
-    return view('layouts.admin_layout');
-});
-
-
-// Route::get('/manage_courses', function () {
-//     return view('admin.manage_courses');
-// });
-
-
-// Route::get('/edit_courses', function () {
-//     return view('admin.edit_courses');
-// });
-
-
-
-
-
-
-// Route::get('/edit_categories', function () {
-//     return view('admin.edit_categories');
-// });
-
-
-Route::get('/manage_users', function () {
-    return view('admin.manage_users');
-});
-
-
-// Route::get('/manage_lessons', function () {
-//     return view('admin.manage_lessons');
-// });
-
-// Route::get('/edit_lessons', function () {
-//     return view('admin.edit_lessons');
-// });
-
-
-// Route::get('/manage_admins', function () {
-//     return view('admin.manage_admins');
-// });
-
-
-// Route::get('/edit_admins', function () {
-//     return view('admin.edit_admins');
-// });
-
-
-// Route::get('/manage_units', function () {
-//     return view('admin.manage_units');
-// });
-
-
-// Route::get('/edit_units', function () {
-//     return view('admin.edit_units');
-// });
-
-
-Route::get('/manage_orders', function () {
-    return view('admin.manage_orders');
-});
-Route::get('/testing', function () {
-    return view('public_website.testinglogIn');
-});
 
 
 
 // Categories Routes
-Route::get('/categories/create','CategoryController@create');
+Route::get('/categories/create','CategoryController@create')->middleware('auth:admin');
 Route::post('/categories','CategoryController@store');
 Route::delete('/categories/{category}','CategoryController@destroy');
-Route::get('/categories/{category}/edit','CategoryController@edit');
+Route::get('/categories/{category}/edit','CategoryController@edit')->middleware('auth:admin');
 Route::put('/categories/{category}','CategoryController@update');
 // Homepage Route
 Route::get('/','CategoryController@index')->name('homepage');
@@ -123,17 +30,17 @@ Route::get('/categories/{category}','CategoryController@show');
 
 
 // Courses Routes
-Route::get('/courses/create','CourseController@create');
+Route::get('/courses/create','CourseController@create')->middleware('auth:admin');
 Route::post('/courses','CourseController@store');
 Route::delete('/courses/{course}','CourseController@destroy');
-Route::get('/courses/{course}/edit','CourseController@edit');
+Route::get('/courses/{course}/edit','CourseController@edit')->middleware('auth:admin');
 Route::put('/courses/{course}','CourseController@update');
 //single course
 Route::get('/courses/{course}','CourseController@show');
 //course add to cart 
 Route::post('/addtocart/{course}','CourseController@AddToCart')->name('course.addtocart');
 //go to cart
-Route::get('/cart' , 'CourseController@getCart');
+Route::get('/cart' , 'CourseController@getCart')->name('cart');
 //delete from cart
 Route::get('cart/{key}/delete', 'CourseController@deleteFromCart')->name('deleteFromCart');
 //go to checkout
@@ -148,20 +55,20 @@ Route::get('/search','CourseController@search');
 
 
 // Units Routes
-Route::get('/units/create','UnitController@create');
-Route::post('/units','UnitController@store');
-Route::delete('/units/{unit}','UnitController@destroy');
-Route::get('/units/{unit}/edit','UnitController@edit');
-Route::put('/units/{unit}','UnitController@update');
+Route::get('/units/create','UnitController@create')->middleware('auth:admin');
+Route::post('/units','UnitController@store')->middleware('auth:admin');
+Route::delete('/units/{unit}','UnitController@destroy')->middleware('auth:admin');
+Route::get('/units/{unit}/edit','UnitController@edit')->middleware('auth:admin');
+Route::put('/units/{unit}','UnitController@update')->middleware('auth:admin');
 
 
 
 // Lessons Routes
-Route::get('/lessons/create','LessonController@create');
-Route::post('/lessons','LessonController@store');
-Route::delete('/lessons/{lesson}','LessonController@destroy');
-Route::get('/lessons/{lesson}/edit','LessonController@edit');
-Route::put('/lessons/{lesson}','LessonController@update');
+Route::get('/lessons/create','LessonController@create')->middleware('auth:admin');
+Route::post('/lessons','LessonController@store')->middleware('auth:admin');
+Route::delete('/lessons/{lesson}','LessonController@destroy')->middleware('auth:admin');
+Route::get('/lessons/{lesson}/edit','LessonController@edit')->middleware('auth:admin');
+Route::put('/lessons/{lesson}','LessonController@update')->middleware('auth:admin');
 
 
 
@@ -176,11 +83,11 @@ Route::put('/admins/{admin}','AdminController@update');
 
 
 // Users Routes
-Route::get('/users/create','UserController@create')->name('user.dashboard');
+Route::get('/users/create','UserController@create')->name('user.dashboard')->middleware('auth:admin');
 Route::post('/users','UserController@store');
-Route::delete('/users/{user}','UserController@destroy');
-Route::get('/users/{user}/edit','UserController@edit');
-Route::put('/users/{user}','UserController@update');
+Route::delete('/users/{user}','UserController@destroy')->middleware('auth:admin');
+Route::get('/users/{user}/edit','UserController@edit')->middleware('auth:admin');
+Route::put('/users/{user}','UserController@update')->middleware('auth:admin');
 
 
 
@@ -202,7 +109,7 @@ Route::prefix('admin')->group(function(){
 
     Route::get('/login' , 'Auth\AdminLoginController@showLoginForm')->name('admin.login');
     Route::post('/logout' , 'Auth\AdminLoginController@adminLogout')->name('admin.logout');
-    Route::post('/login' , 'Auth\AdminLoginController@login')->name('admin.login.submit');
+    Route::post('/login' , 'Auth\AdminLoginController@admin_login')->name('admin.login.submit');
     // Route::get("/", "AdminController@show")->name('admin.dashboard');
 
     //password reset routes
@@ -223,7 +130,7 @@ Route::prefix('admin')->group(function(){
 
 // Orders Routes
 Route::get('/orders','OrderController@index')->name('order.index')->middleware('auth');
-Route::get('/orders/create','OrderController@create')->name('order.dashboard');
+Route::get('/orders/create','OrderController@create')->name('order.dashboard')->middleware('auth:admin');
